integrated cordawyns cleanup and restructured to be nicer to gem
