+++ lib/8583/bitmap.rb
