+++ lib/8583/exception.rb
