+++ lib/8583/field.rb
