+++ lib/8583/fields.rb
