+++ lib/8583/util.rb
