+++ lib/iso8583.rb
@@ -1,12 +1,5 @@
 $:.unshift(File.dirname(__FILE__)) unless
   $:.include?(File.dirname(__FILE__)) || $:.include?(File.expand_path(File.dirname(__FILE__)))
 
-module ISO8583
-  require "field"
-  require "codec"
-  require "fields"
-  require "exception"
-  require "bitmap"
-  require "message"
-  require "util"
-end
+
+require '8583'
